backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pyodbc==5.0.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
typing-extensions==4.8.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
